Better support typed properties
Upgrades the underlying squirrelphp/strings component. This
removes compatibility to diverging property paths defined
in the form, we now only support direct properties which are
mapped 1:1 in the form.

Note that if filter annotations are used without direct mappings,
this might lead to errors and/or not work anymore.

The PHP 8 support is preliminary, as it was not tested yet.